Separated ball projection code into separate file
